Fix unit tests - Handle a recording which was stopped before any media was recorded
